refactor: refactored the file structure
